Refactor: 폴더구조 개편
각각 관심사에 맞게 파일 구조를 개편하였습니다.

1. 페이지만 담당한 pages폴더
2. 기능들을 관심사별로 분리하면 features폴더
3. 전역에 영향을 끼치는 기능을 모아놓은 apps폴더
4. 서비스로직이 존재하는 services폴더
