--- src/NotFoundPage.jsx
@@ -1,3 +1,0 @@
-export default function NotFoundPage() {
-  return (<p>Not Found</p>);
-}
