--- src/NotFoundPage.test.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { render } from '@testing-library/react';
-
-import NotFoundPage from './NotFoundPage';
-
-test('NotFoundPage', () => {
-  render(<NotFoundPage />);
-});
