+++ src/features/region/RegionsContainer.jsx
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from './actions';
+} from '../../apps/actions';
 
-import { get } from './utils';
+import { get } from '../../apps/utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
