+++ src/pages/About.test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { render } from '@testing-library/react';
+
+import About from './About';
+
+test('About', () => {
+  render(<About />);
+});
