+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 
-export default function HomePage() {
+export default function Home() {
   return (
     <div>
       <h1>Home</h1>
