+++ src/pages/NotFound.jsx
@@ -1,0 +1,3 @@
+export default function NotFound() {
+  return (<p>Not Found</p>);
+}
